% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data Simulation.R
\name{generate_individual}
\alias{generate_individual}
\title{generate_individual}
\usage{
generate_individual(
  id = 1,
  T,
  P,
  sigma_residual,
  sigma_randint,
  main_rand = 0.5,
  rho,
  beta_logit,
  model,
  beta,
  theta1
)
}
\arguments{
\item{id}{the assigned id for this specific individual}

\item{T}{the number of observations for this individual}

\item{P}{the number of measurements or predictors}

\item{sigma_residual}{the noise added to the moderator formula when generate pesudo outcome. This noise is
generated using ARIMA}

\item{sigma_randint}{the noise added to the moderator formula when generate pesudo outcome.}

\item{main_rand}{the noise added to the control formula when generate pesudo outcome. This noise is generated using ARIMA}

\item{rho}{this value is used to simulate state or predictors. We also use ARIMA when generating predictors, and
rho defines how much of the past values influence the current value in an autoregressive model}

\item{beta_logit}{the true value of beta when simulate the probability of assign treatment $p(A_t = 1|H_t)$}

\item{model}{the true model for moderator when generate pesudo outcomes}

\item{beta}{the true coefficients for moderator}

\item{theta1}{the coefficients for the working model of control variables. For now, we just simply assume the working model
contains the all predictors with coefficients equal to theta1}
}
\value{
A data frame containing the generated states $S_t$, action $A_t$, outcomes $Y_{t+1}$, treatment effects, and noise components for a single participant.
}
\description{
simulates data for an individual participant
}
\examples{
generate_individual(id = 1, T = 50, P = 50, sigma_residual = 1.5, sigma_randint = 1.5, main_rand = 2, rho = 0.7,
beta_logit = c(-1, 1.6 * rep(1/50, 50)), model = ~ state1 + state2 + state3 + state4,
beta = matrix(c(-0.2, 0.8, 0.3, 0.7, 0.3),ncol = 1),
theta1 = 0.8
)
}
