% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{run_simple_lasso_test}
\alias{run_simple_lasso_test}
\title{run_simple_lasso_test}
\usage{
run_simple_lasso_test(
  venv,
  n = 200L,
  p = 20L,
  s = 5L,
  sigma = 1,
  rho = 0.3,
  signal = 3,
  lambda_frac = 1
)
}
\description{
Test if the virtual environment has been successfully loaded
}
