% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomLASSO_subgrad_Py.R
\name{variable_selection_PY_penal_int}
\alias{variable_selection_PY_penal_int}
\title{variable_selection_PY_penal_int}
\usage{
variable_selection_PY_penal_int(
  data,
  ID,
  moderator_formula,
  lam = NULL,
  noise_scale = NULL,
  splitrat = 0.8,
  virtualenv_path = "",
  ridge_term = 0,
  beta = NULL
)
}
\arguments{
\item{data}{the output of pesudo_outcomecal function}

\item{ID}{the name of column where participants' ID are stored}

\item{moderator_formula}{determines the formula for the f(St)T*beta function}

\item{lam}{the value of penalty term of randomized LASSO. If it is not provided, the default value will be used}

\item{noise_scale}{Scale of Gaussian noise added to objective. Default is \eqn{\sqrt{\frac{(1 - \rho)}{\rho}*number\ of\ rows}*sd(y)} where \eqn{\rho} is the split rate.
The random noises, \eqn{\omega}, are iid drawn from normal distribution with standard deviation noise_scale}

\item{splitrat}{this value is corresponding to the data splitting rate. Details can read "Exact Selective Inference with Randomization" page 15 equation (10).
This value will be used only when user doesn't provide the noise_scale.}

\item{virtualenv_path}{Python virtual environment path}
}
\value{
A table with the selected variables is returned. The returned table contains 
Selected variables that remain in the model after penalization, Non-selected variables that were excluded from the model,
Number of unique subjects in the dataset, Random noise added for each variables, Regularization parameter (lambda) used in the penalization, Subgradient of unselected variables, 
OMEGA: (noise_scale)^2/4. Divide 4 because in theoretical calculation -1/2 is hidden inside penalty and random noise, sign_soln: Signs of the estimated coefficients for selected variable,
Estimated coefficients of selected variables from the penalized regression, True beta projected on space of post-selection predictors, Boolean vector indicating which variables were selected.
}
\description{
This version allow the program penalize the intercept
}
\examples{

var_selection_python = variable_selection_PY_penal_int(data = pseudo_outcome_CVlasso,ID,
my_formula, lam = NULL, noise_scale = NULL, splitrat = 0.8,
virtualenv_path = "fakepath/wcls", ridge_term = 0, beta = NULL)

}
