% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UI_function_LASSO.R
\name{DR_WCLS_LASSO}
\alias{DR_WCLS_LASSO}
\title{DR_WCLS_LASSO}
\usage{
DR_WCLS_LASSO(
  data,
  fold,
  ID,
  time,
  Ht,
  St,
  At,
  outcome,
  method_pesu,
  lam = NULL,
  noise_scale = NULL,
  splitrat = 0.8,
  virtualenv_path,
  beta = NULL,
  level = 0.9,
  core_num = NULL
)
}
\arguments{
\item{data}{raw data without pseudo-outcome, ptSt}

\item{fold}{# of folds hope to split when generating pesudo outcome}

\item{ID}{the name of column where participants' ID are stored}

\item{time}{the name of column where time in study are stored}

\item{Ht}{a vector that contains column names of control variables}

\item{St}{a vector that contains column names of moderator variables; St should be a subset of Ht}

\item{outcome}{column names of outcome variable}

\item{method_pesu}{the machines learning method used when generate estimates of the nuisance parameters,
and those values will be used to calculate the pseudo outcome. The available machine learning algorithms include
cross validation LASSO ("CVLASSO"), Random Forest ("RandomForest"), and gradient boosting ("GradientBoosting").}

\item{lam}{the value of penalty term of randomized LASSO. If it is not provided, the default value will be used}

\item{noise_scale}{Scale of Gaussian noise added to objective. Default is \eqn{\sqrt{\frac{(1 - \rho)}{\rho}*number\ of\ rows}*sd(y)} where \eqn{\rho} is the split rate.
The random noises, \eqn{\omega}, are iid drawn from normal distribution with standard deviation noise_scale}

\item{splitrat}{this value is corresponding to the data splitting rate. Details can read "Exact Selective Inference with Randomization" page 15 equation (10).
This value will be used only when user doesn't provide the noise_scale.}

\item{virtualenv_path}{Python virtual environment path}

\item{beta}{the true coefficient value (if simulation is conducted)}

\item{level}{the CI significant level}

\item{core_num}{the number of cores will be used for parallel calculation}
}
\value{
A table with the selected variables is returned. The returned table contains GEE estimate,
the post selection true value (if simulation is conducted; Otherwise, NA is provided.)
the p value,
the confidence interval,
the true corresponding pivot value for the lower bound and the upper bound.
}
\description{
A function for basic LASSO DR-WCLS
}
\examples{

UI_return = DR_WCLS_LASSO(data = data, fold = 5, ID = "id", time = "decision_point",
Ht = Ht, St = St, At = "action", outcome = "outcome", method_pesu = "CVLASSO",
lam = NULL, noise_scale = NULL, splitrat = 0.8,
virtualenv_path = "path to selective-inference folder/env3",
beta =  c(-0.2, 0.8, 0.3, 0.7, 0.3, rep(0, 21)), level = 0.9, core_num = 3)

}
